hiver_core::aware

Trait BeanFactoryAware

pub trait BeanFactoryAware {
    // Required method
    fn set_bean_factory(&mut self, container: &Container);
}

Trait for beans that need a reference to the IoC container. 需要IoC容器引用的Bean trait。

Equivalent to Spring’s BeanFactoryAware. 等价于 Spring 的 BeanFactoryAware。

This is useful when a bean needs to look up other beans programmatically. 当Bean需要以编程方式查找其他Bean时，这很有用。

Example / 示例

use hiver_core::aware::BeanFactoryAware;
use hiver_core::Container;

struct RouterBean {
    container: Option<Container>,
}

impl BeanFactoryAware for RouterBean {
    fn set_bean_factory(&mut self, container: &Container) {
        self.container = Some(container.clone());
    }
}

Required Methods

fn set_bean_factory(&mut self, container: &Container)

Supply the owning Container to this bean. 将拥有的 Container 提供给此Bean。