hiver_core::lifecycle

Trait InitializingBean

pub trait InitializingBean {
    // Required method
    fn after_properties_set(&mut self) -> Result<()>;
}

Trait for beans that need initialization after property injection. 在属性注入后需要初始化的Bean trait。

Equivalent to Spring’s InitializingBean interface or @PostConstruct. 等价于 Spring 的 InitializingBean 接口或 @PostConstruct。

When it fires / 何时触发

Called after all properties have been supplied to the bean. 在所有属性注入到Bean之后调用。

Example / 示例

use hiver_core::lifecycle::InitializingBean;

struct MyService
{
    initialized: bool,
}

impl InitializingBean for MyService
{
    fn after_properties_set(&mut self) -> hiver_core::lifecycle::Result<()>
    {
        self.initialized = true;
        Ok(())
    }
}

Required Methods

fn after_properties_set(&mut self) -> Result<()>

Called after all bean properties have been set. 在所有Bean属性设置完毕后调用。

Equivalent to @PostConstruct. 等价于 @PostConstruct。